+++ .eslintrc.js
@@ -99,6 +99,7 @@ module.exports = {
     'camelcase': 'off',
     'no-control-regex': 'off',
     'no-await-in-loop': 'off',
+    'import/no-extraneous-dependencies': 'off',
   },
   'globals': {
     'window': true,