+++ packages/mp-webpack-plugin/src/tmpl/custom-component.tmpl.js
@@ -35,6 +35,7 @@ Component({
     },
     options: {
         addGlobalClass: true, // 开启全局样式
+        virtualHost: true, // 开启虚拟化 host
     },
     attached() {
         const nodeId = this.dataset.privateNodeId
@@ -53,8 +54,8 @@ Component({
 
         // 监听全局事件
         this.onSelfNodeUpdate = tool.throttle(this.onSelfNodeUpdate.bind(this))
-        this.domNode.$$clearEvent('$$domNodeUpdate')
-        this.domNode.addEventListener('$$domNodeUpdate', this.onSelfNodeUpdate)
+        this.domNode.$$clearEvent('$$domNodeUpdate', {$$namespace: 'proxy'})
+        this.domNode.addEventListener('$$domNodeUpdate', this.onSelfNodeUpdate, {$$namespace: 'proxy'})
 
         // 监听自定义组件事件
         const {events = []} = this.compConfig